refactor(core): the dispatch moves to where both callers can reach it — 308 verb specs + 109 admission + 66 constraint unchanged, 271 cargo tests, zero behaviour
Phase 3, slice 14. The Command enum, the parse/plan/dispatch/respond
surface and the {ok, error:{code,message,hint}} envelope lift out of the
CLI binary into crates/wealth-core/src/command.rs, so the desktop
shell's one Tauri command and the CLI bin will run THE SAME exhaustive
match — the property that a new verb cannot exist without the compiler
demanding its dispatch arm. The bin shrinks from 466 lines to 159:
argument parsing, stdio, exit codes — a caller, not a surface.

Zero behaviour, measured not asserted: every suite count identical
before and after with no test file touched; eleven CLI invocations
captured byte-for-byte across stdout, stderr and exit codes — success,
hinted refusal, constraint refusal, serde refusing an unknown verb
before any connection exists, sub-penny money, plan with and without a
database, faults. The structural mutation adds an unarmed enum variant
and the build fails on the non-exhaustive match — and the lift makes
that guarantee STRONGER: it used to fire only under the cli feature
flag; it now fires on a plain library build, which is the build the
shell will use.

The public surface is seven items; parse deliberately takes text rather
than a Value, because from_str reports line and column and from_value
cannot — one mistake must not have two different messages on the wire.

Named for the next PR, not smuggled into this one: the ten pre-existing
verb-lane reds trace to a LIVE cloud defect — restore materialises
backups through jsonb_populate_recordset, which emits explicit NULLs
for keys a backup omits, and needs_review is now NOT NULL, so restoring
any backup written before that column existed would fail in production.
